+++ alexflipnote/__init__.py
@@ -1,4 +1,7 @@
-from .classes import Colour, Icon, Image, Steam
+from .classes import Colour, Icon, Image
 from .client import BadRequest, Client, InternalServerError, NotFound
+from .errors import BadRequest, HTTPException, InternalServerError, NotFound
 
-__version__ = '1.5.5'
+__license__ = 'MIT'
+__author__ = 'Soheab_'
+__version__ = '1.6.0'